refactor: 참여중인 방 조회 기능 수정
