+++ backend/src/main/java/corea/matching/repository/ParticipationRepository.java
@@ -10,4 +10,6 @@ public interface ParticipationRepository extends JpaRepository<Participation, Lo
     List<Participation> findAllByRoomId(long roomId);
 
     List<Participation> findAllByMemberId(long memberId);
+
+    boolean existsByRoomIdAndMemberId(long roomId, long memberId);
 }
